requirements.txt
#
# This file is autogenerated by pip-compile with python 3.8
# To update, run:
#
#    pip-compile
#
mysql-connector-python
requests
aiohttp==3.8.1
    # via -r requirements.in
aiomysql==0.1.1
    # via -r requirements.in
aiosignal==1.2.0
    # via aiohttp
alembic==1.8.0
    # via -r requirements.in
anyio==3.6.1
    # via starlette
async-timeout==4.0.2
    # via aiohttp
attrs==21.4.0
    # via aiohttp
bcrypt==3.2.2
    # via -r requirements.in
cffi==1.15.0
    # via
    #   bcrypt
    #   cryptography
charset-normalizer==2.0.12
    # via aiohttp
click==8.1.3
    # via uvicorn
cryptography==37.0.2
    # via -r requirements.in
databases==0.5.5
    # via
    #   -r requirements.in
    #   ormar
decorator==5.1.1
    # via validators
dnspython==2.2.1
    # via email-validator
ecdsa==0.17.0
    # via python-jose
email-validator==1.2.1
    # via -r requirements.in
et-xmlfile==1.1.0
    # via openpyxl
fastapi==0.78.0
    # via -r requirements.in
frozenlist==1.3.0
    # via
    #   aiohttp
    #   aiosignal
greenlet==1.1.2
    # via sqlalchemy
gunicorn==20.1.0
    # via -r requirements.in
h11==0.13.0
    # via uvicorn
idna==3.3
    # via
    #   anyio
    #   email-validator
    #   yarl
mako==1.2.0
    # via alembic
markupsafe==2.1.1
    # via mako
multidict==6.0.2
    # via
    #   aiohttp
    #   yarl
numpy==1.22.4
    # via
    #   opencv-python
    #   pandas
    #   scipy
opencv-python==4.6.0.66
    # via -r requirements.in
openpyxl==3.0.10
    # via -r requirements.in
ormar==0.11.2
    # via -r requirements.in
paho-mqtt==1.6.1
    # via -r requirements.in
pandas==1.4.2
    # via -r requirements.in
passlib==1.7.4
    # via -r requirements.in
pillow==9.1.1
    # via -r requirements.in
pyasn1==0.4.8
    # via
    #   python-jose
    #   rsa
pycparser==2.21
    # via cffi
pydantic==1.8.2
    # via
    #   fastapi
    #   ormar
pymysql==1.0.2
    # via
    #   -r requirements.in
    #   aiomysql
python-dateutil==2.8.2
    # via pandas
python-dotenv==0.20.0
    # via -r requirements.in
python-jose==3.3.0
    # via -r requirements.in
python-multipart==0.0.5
    # via -r requirements.in
pytz==2022.1
    # via pandas
qrcode==7.3.1
    # via -r requirements.in
rsa==4.8
    # via python-jose
scipy==1.8.1
    # via -r requirements.in
six==1.16.0
    # via
    #   ecdsa
    #   python-dateutil
    #   python-multipart
sniffio==1.2.0
    # via anyio
sqlalchemy==1.4.31
    # via
    #   -r requirements.in
    #   alembic
    #   databases
    #   ormar
starlette==0.19.1
    # via fastapi
tqdm==4.64.0
    # via -r requirements.in
typing-extensions==4.2.0
    # via pydantic
uvicorn==0.18.2
    # via -r requirements.in
validators==0.20.0
    # via -r requirements.in
xlrd==2.0.1
    # via -r requirements.in
yarl==1.7.2
    # via aiohttp

# The following packages are considered to be unsafe in a requirements file:
# setuptools
